Fixed reading values from nametrees; Added support for named destinations; Fixed a memory leak in PdfDestination::GetPage; Pages do not know their page number inside of the document; Fixed reading PdfActions from PDF files; Fixed some warnings;
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@309 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfImage.cpp
@@ -138,7 +138,7 @@ void PdfImage::LoadFromFile( const char* pszFilename )
         RAISE_ERROR( ePdfError_OutOfMemory );
     }
 
-    if( fread( szBuffer, sizeof( char ), lLen, hInfile ) != lLen )
+    if( static_cast<long>(fread( szBuffer, sizeof( char ), lLen, hInfile )) != lLen )
     {
         fclose( hInfile );
         free( szBuffer );
